+++ tests/test_integration/test_openapi/test_schema.py
@@ -111,7 +111,6 @@ class _DjangoSessionAuth:
 
 
 @schema.parametrize()
-@pytest.mark.django_db(transaction=True)
 @h_settings(max_examples=_MAX_EXAMPLES)
 @pytest.mark.timeout(0)
 def test_schemathesis(